Move peggo directory to lib
This more accurately represents what is in the directory
