+++ src/CMakeLists.txt
@@ -1,1 +1,1 @@
-add_subdirectory(peggo)
+add_subdirectory(lib)
