+++ src/lib/common.c
