+++ src/lib/common.h
